+++ frontend/src/main.js
@@ -5,7 +5,6 @@ import './styles/tailwind.css'; // Tailwind CSS
 import BootstrapVue3 from 'bootstrap-vue-3'; // BootstrapVue
 import { createPinia } from 'pinia'; // Pinia
 import router from './router'; // Vue Router
-import router from "@/router/router.js";
 
 import 'bootstrap/dist/css/bootstrap.css';
 import 'bootstrap-vue-3/dist/bootstrap-vue-3.css';
